Maintain backwards compatiblity with Capistrano 2
